.zshrc.local: Increase zsh history
Increase the number of events stored in the history list, see: zshparam(1).

HISTFILE is set in .zshrc (default in grml-config). If unset, the
history is not saved.

HISTSIZE is number of events in the internal history list (call it
memory). Only if SAVEHIST is set, the history is saved (if HISTFILE is
set).