.gitconfig: Add some more git tweaks
To use vimdiff as a git mergetool the [merge] and [mergetool] config
section is needed, see: https://www.rosipov.com/blog/use-vimdiff-as-git-mergetool/

diff.noprefix is nice to remove the a/ and b/ prefix when the diffs are
shown.

log.abbrevCommit is nice because it only shows the short commit id in
git log and git show.