.zshrc.local: Convert lsshot + cpshot aliases to functions
Aliases can not be used with parameters, to be able to customize the
output of lsshot and cpshot, function is needed.

By default lsshot shows the last 10 modified screenshots but by using
some other value as parameter the number of files can be changed.

And cpshot by default only copies the last modified screenshot.
By using a parameters the number of latest screenshots can be set.