.config/starship.toml: Basic starship configuration
Starship is cross-shell prompt. It is can be installed with:

  curl -fsSL https://starship.rs/install.sh | sudo bash

And enabled for zsh via:

  # ~/.zshrc

  eval "$(starship init zsh)"

By default the commit hash is only showed for detached HEAD,
but I like to see the commit hash all the time.

Also the time module is disabled by default, but I like to see the date
and time all the time.

Currently the shell looks like this:

  configs on  master (434499b) [$+⇡] took 3m51s
  ✦ at 2021-02-17 19:49:09 +01:00 ❯

The username is hostname currently not shown and the path is truncated
too much. This should be fixed.