.config/starship.toml: Configure directory module
By default the directory module shows the path to the current directory,
truncated to three parent folders. It also truncated to the root of
the current git repo.

I want to avoid truncation, even in a git repo. Start with length 5.
This might be changed in the future. But in case it is truncated I need
to know and the "../" is prefixed.

It also turned out that the username + hostname are only shown in
certain conditions. This is fine for me.