.vimrc.local: Enabled vim-script / vim-addon vimplate
vimplate is a template system for vim.
This is a sophisticated replacement for the imap ,jst which added a
journal template.

With the Template Toolkit, a template processing system, this is way
more powerful. This addon depends on 'libtemplate-perl'.

To add the template run :Vimplate journal. Then 'Topic', 'Why' and
'What' is asked and used in the template.