.config/starship.toml: Configure battery module
The battery module shows how charged the device's battery is and its
current charging status. The module is only visible when the device's
battery is below 10% by default.

The default icons do not exist in the Fira Code font set, so I used
different fonts which do exist in Fira Code.

Also changed the threshold to show the battery symbol below 20%.